Renamed database.rb to database_adapter.rb
