+++ lib/wgit.rb
@@ -9,7 +9,7 @@ require_relative 'wgit/document'
 require_relative 'wgit/document_extractors'
 require_relative 'wgit/crawler'
 require_relative 'wgit/database/model'
-require_relative 'wgit/database/database'
+require_relative 'wgit/database/database_adapter'
 require_relative 'wgit/database/mongo_db'
 require_relative 'wgit/robots_parser'
 require_relative 'wgit/indexer'
