+++ lib/wgit/database/database_adapter.rb
