+++ load.rb
@@ -12,7 +12,7 @@ load 'lib/wgit/document.rb'
 load 'lib/wgit/document_extractors.rb'
 load 'lib/wgit/crawler.rb'
 load 'lib/wgit/database/model.rb'
-load 'lib/wgit/database/database.rb'
+load 'lib/wgit/database/database_adapter.rb'
 load 'lib/wgit/database/mongo_db.rb'
 load 'lib/wgit/robots_parser.rb'
 load 'lib/wgit/indexer.rb'
